+++ code/renderers/react/src/componentManifest/fixtures.ts
@@ -110,3 +110,84 @@ export const fsMocks = {
           </header>
       );`,
 };
+
+export const indexJson = {
+  v: 5,
+  entries: {
+    'example-button--primary': {
+      type: 'story',
+      subtype: 'story',
+      id: 'example-button--primary',
+      name: 'Primary',
+      title: 'Example/Button',
+      importPath: './src/stories/Button.stories.ts',
+      componentPath: './src/stories/Button.tsx',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      exportName: 'Primary',
+    },
+    'example-button--secondary': {
+      type: 'story',
+      subtype: 'story',
+      id: 'example-button--secondary',
+      name: 'Secondary',
+      title: 'Example/Button',
+      importPath: './src/stories/Button.stories.ts',
+      componentPath: './src/stories/Button.tsx',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      exportName: 'Secondary',
+    },
+    'example-button--large': {
+      type: 'story',
+      subtype: 'story',
+      id: 'example-button--large',
+      name: 'Large',
+      title: 'Example/Button',
+      importPath: './src/stories/Button.stories.ts',
+      componentPath: './src/stories/Button.tsx',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      exportName: 'Large',
+    },
+    'example-button--small': {
+      type: 'story',
+      subtype: 'story',
+      id: 'example-button--small',
+      name: 'Small',
+      title: 'Example/Button',
+      importPath: './src/stories/Button.stories.ts',
+      componentPath: './src/stories/Button.tsx',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      exportName: 'Small',
+    },
+    'example-header--docs': {
+      id: 'example-header--docs',
+      title: 'Example/Header',
+      name: 'Docs',
+      importPath: './src/stories/Header.stories.ts',
+      type: 'docs',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      storiesImports: [],
+    },
+    'example-header--logged-in': {
+      type: 'story',
+      subtype: 'story',
+      id: 'example-header--logged-in',
+      name: 'Logged In',
+      title: 'Example/Header',
+      importPath: './src/stories/Header.stories.ts',
+      componentPath: './src/stories/Header.tsx',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      exportName: 'LoggedIn',
+    },
+    'example-header--logged-out': {
+      type: 'story',
+      subtype: 'story',
+      id: 'example-header--logged-out',
+      name: 'Logged Out',
+      title: 'Example/Header',
+      importPath: './src/stories/Header.stories.ts',
+      componentPath: './src/stories/Header.tsx',
+      tags: ['dev', 'test', 'vitest', 'autodocs'],
+      exportName: 'LoggedOut',
+    },
+  },
+};
